+++ src/errorOut.ts
@@ -1,8 +1,12 @@
 import { logError } from "./logError";
 import { FirebaseError } from "./error";
+import { setupLoggers } from "./utils";
 
 /**
- * Errors out by calling `process.exit` with an exit code of 2.
+ * Sets up the process to be exited with an error code specified by the error
+ * or a default of `2`.
+ * Does *not* call process.exit. End the process by returning from the
+ * callsite.
  * @param error an Error to be logged.
  */
 export function errorOut(error: Error): void {
@@ -16,9 +20,9 @@ export function errorOut(error: Error): void {
     });
   }
 
+  // In case we've not set them up yet, set up loggers.
+  setupLoggers();
+
   logError(fbError);
   process.exitCode = fbError.exit || 2;
-  setTimeout(() => {
-    process.exit();
-  }, 250);
 }